+++ src/components/manage/Toolbar/PersonalTools.test.jsx
@@ -37,7 +37,48 @@ describe('Toolbar Personal Tools component', () => {
     const component = renderer.create(
       <Provider store={store}>
         <PluggablesProvider>
-          <MemoryRouter>
+          <MemoryRouter initialEntries={[{ pathname: '' }]}>
+            <PersonalTools
+              loadComponent={() => {}}
+              theToolbar={{
+                current: { getBoundingClientRect: () => ({ width: '320' }) },
+              }}
+            />
+          </MemoryRouter>
+        </PluggablesProvider>
+      </Provider>,
+    );
+    const json = component.toJSON();
+    expect(json).toMatchSnapshot();
+  });
+
+  it('renders an Toolbar Personal Tools component - in a non-root path', () => {
+    const store = mockStore({
+      users: {
+        user: {
+          fullname: 'admin',
+          email: 'admin@plone.org',
+          roles: ['Manager'],
+        },
+      },
+      userSession: {
+        token: jwt.sign({ sub: 'admin' }, 'secret'),
+      },
+      content: {
+        data: {
+          '@type': 'Folder',
+          is_folderish: true,
+        },
+      },
+      intl: {
+        locale: 'en',
+        messages: {},
+      },
+    });
+    const component = renderer.create(
+      <Provider store={store}>
+        <PluggablesProvider>
+          <MemoryRouter initialEntries={[{ pathname: '/en/blog' }]}>
             <PersonalTools
               loadComponent={() => {}}
               theToolbar={{