+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Fix unlock after changing the id and saving a page @reebalazs
 - Group routes so React does not see them as a different Route and triggers a full remount. This is specially important in `Contents` @sneridagh
 - Add default to `null` for `token` prop in `Navigation` component. This prevents the component to shoot an extra call when the logout happens @sneridagh
+- Fix a double slash present in the `PersonalTools` component @sneridagh
 
 ### Internal