+++ CHANGELOG.md
@@ -70,6 +70,7 @@
 
 ### Internal
 
+- Add some internal block utilities `getBlockInitialDataDefaults` and `isBlockResetable` @sneridagh
 - Run yarn deduplicate on dependencies. @davisagli
 
 ### Documentation